antispoof/requirements.txt
clarifai-grpc>=10.0.0
python-dotenv>=1.0.0

